#rstats R Day, Brazil. May, 2018
R Day - National Meeting of R users is an event to congregate users and
developers of R in an environment for the exchange of experiences, learning
and contacts.

R Day will happen in UFPR, Curitiba (PR), Brazil, on May 22, 2018. The goal of R Day is to bring together an active and large, though sparse, community of R users in Brazil.

This will be the first event of R in Brazil officially endorsed by The R Foundation. (359)